start db connection after getting GET\POST requests
